requirements.txt
pydantic-ai
beautifulsoup4
nest_asyncio
httpx
python-dotenv
requests